requirements.txt
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.3.0
colorama==0.4.6
dash==3.3.0
dash-bootstrap-components==2.0.4
dash_ag_grid==32.3.2
duckdb==1.4.2
Flask==3.1.2
idna==3.7
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.11.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==25.0
pandas==2.2.2
plotly==6.4.0
polygon-api-client==1.15.3
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.1
requests==2.32.2
retrying==1.4.2
setuptools==80.9.0
six==1.16.0
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.2.1
websockets==14.2
Werkzeug==3.1.3
zipp==3.23.0
